requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.12
aniso8601==8.0.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clyent==1.2.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
cryptography==2.9.2
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flask-restplus==0.13.0
flask-restx==0.2.0
future==0.18.2
ga==0.1.1.dev0
glob2==0.7
google-auth==1.18.0
google-pasta==0.2.0
grpcio==1.29.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel==5.3.0
ipython @ file:///C:/ci/ipython_1593433781269/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.1.5
Keras-Preprocessing==1.1.2
libarchive-c==2.9
Markdown==3.2.2
MarkupSafe==1.1.1
menuinst==1.4.16
mistune==0.8.4
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
olefile==0.46
opt-einsum==3.2.1
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
pipreqs==0.4.10
pkginfo==1.5.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
rsa==4.6
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sudo==1.0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1593446365756/work
traitlets==4.3.3
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xmltodict==0.12.0
yarg==0.1.9
zipp==3.1.0
